requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
    #   django-cors-headers
astroid==3.3.5
    # via
    #   -r requirements/test.in
    #   pylint
attrs==24.2.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
cachetools==5.5.0
    # via tox
certifi==2024.8.30
    # via
    #   -r requirements/base.txt
    #   elastic-transport
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
chardet==5.2.0
    # via
    #   diff-cover
    #   tox
charset-normalizer==3.4.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.8.0
    # via -r requirements/test.in
colorama==0.4.6
    # via tox
coverage[toml]==7.6.4
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==43.0.3
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.7.2
    # via -r requirements/test.in
diff-cover==9.2.0
    # via -r requirements/test.in
dill==0.3.9
    # via pylint
distlib==0.3.9
    # via virtualenv
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django-cors-headers
    #   django-crum
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   drf-spectacular
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
django-cors-headers==4.6.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-elasticsearch-dsl==8.0
    # via -r requirements/base.txt
django-waffle==4.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   drf-spectacular
    #   edx-drf-extensions
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
drf-spectacular==0.27.2
    # via -r requirements/base.txt
edx-django-release-util==1.4.0
    # via -r requirements/base.txt
edx-django-utils==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==10.5.0
    # via -r requirements/base.txt
edx-opaque-keys==2.11.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
elastic-transport==8.15.1
    # via
    #   -r requirements/base.txt
    #   elasticsearch
elasticsearch==8.15.1
    # via
    #   -r requirements/base.txt
    #   elasticsearch-dsl
elasticsearch-dsl==8.15.4
    # via
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl
factory-boy==3.3.1
    # via -r requirements/test.in
faker==30.8.2
    # via factory-boy
filelock==3.16.1
    # via
    #   tox
    #   virtualenv
gunicorn==23.0.0
    # via -r requirements/base.txt
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via pylint
jinja2==3.1.4
    # via
    #   code-annotations
    #   diff-cover
jsonschema==4.23.0
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
more-itertools==10.5.0
    # via -r requirements/test.in
mysqlclient==2.2.5
    # via -r requirements/base.txt
newrelic==10.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==24.2
    # via
    #   -r requirements/base.txt
    #   gunicorn
    #   pyproject-api
    #   pytest
    #   tox
path==17.0.0
    # via
    #   -r requirements/base.txt
    #   path-py
path-py==12.5.0
    # via -r requirements/base.txt
pbr==6.1.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pep8==1.7.1
    # via -r requirements/test.in
platformdirs==4.3.6
    # via
    #   pylint
    #   tox
    #   virtualenv
pluggy==1.5.0
    # via
    #   diff-cover
    #   pytest
    #   tox
psutil==6.1.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.18.0
    # via diff-cover
pyjwt[crypto]==2.9.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pylint==3.3.1
    # via -r requirements/test.in
pymongo==4.10.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pyproject-api==1.8.0
    # via tox
pytest==8.3.3
    # via
    #   -r requirements/test.in
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.9.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   elasticsearch-dsl
    #   faker
python-slugify==8.0.4
    # via code-annotations
pytz==2024.2
    # via -r requirements/base.txt
pyyaml==6.0.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   drf-spectacular
    #   edx-django-release-util
referencing==0.35.1
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
rpds-py==0.21.0
    # via
    #   -r requirements/base.txt
    #   jsonschema
    #   referencing
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   django-elasticsearch-dsl
    #   edx-django-release-util
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.3.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.2
    # via pylint
tox==4.23.2
    # via -r requirements/test.in
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
    #   elasticsearch-dsl
    #   faker
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   drf-spectacular
urllib3==2.2.3
    # via
    #   -r requirements/base.txt
    #   elastic-transport
    #   requests
virtualenv==20.27.1
    # via tox

# The following packages are considered to be unsafe in a requirements file:
setuptools==75.3.0
    # via -r requirements/base.txt
